organização pastas e componentes do perfil
