+++ src/pages/perfil/perfil.css
@@ -19,3 +19,12 @@
     justify-content: center;
  
 }
+
+.mainPerfil{
+    display: flex;
+    flex-direction: column;
+    max-width: 70%;
+    align-items: center;
+    justify-content: center;
+    margin-left: 300px;
+}
