--- src/pages/perfil/perfilSuperior.css
@@ -1,8 +1,0 @@
-.mainPerfil{
-    display: flex;
-    flex-direction: column;
-    max-width: 70%;
-    align-items: center;
-    justify-content: center;
-    margin-left: 300px;
-}
